+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+### 2012-08-27
+* Added `UserResponseInterface#getRealName()` method, also a new default path `realname`
+  was added, this path holds the real name of user
+* Added `UserResponseInterface#getNickName()` method, also a new default path `nickname`
+  was added, this path holds the nickname of user
+* [BC break] Renamed path `username` to `identifier` to make it more clear that this path should
+  hold the unique user identifier (previously `username`)
+* [BC break] Method `UserResponseInterface#getUsername()` now always returns a real
+  unique user identifier, and uses path `identifier`
+* [BC break] `OAuth1RequestTokenStorageInterface#save()` second param `$token` must
+  now be an array
+
 ### 2012-07-15
 
 * Added `UserResponseInterface#getAccessToken()` and `UserResponseInterface#setAccessToken`